Increase block align size to speed up GC.
This appears to work, but not sure what the ramifications are.

Add NS implementation for estimated load progress

Eval javascript async by default.

Switch default module prefix to so instead of dylib

ZMQ relies on this for creating the path of its shared object lib.

Changes to make tramp able to ssh into windows WSL machines.

Minor changes

Add fringe active face for when the current window is active.